# api/openapi.yaml
openapi: 3.0.3
info:
  title: Photoffice API
  description: |
    REST-API des Photoffice-Backends. Diese Datei ist die Quelle der Wahrheit (API-first):
    Das Backend generiert daraus Java-Interfaces, das Frontend einen Angular-Client.
  version: 0.1.0
servers:
  - url: /api
security:
  - bearerAuth: []
...
paths:
...
  /platform/tenants:
    get:
      tags: [platform]
      operationId: listTenants
      summary: Alle Studios, sortiert nach Name
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TenantResponse'
    post:
      tags: [platform]
      operationId: createTenant
      summary: Neues Studio registrieren
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTenantRequest'
      responses:
        '201':
          description: Studio angelegt
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TenantResponse'
        '400':
          description: Ungültige Eingabe
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '401':
          description: Nicht angemeldet
        '403':
          description: Keine Berechtigung (nur Plattform-Betreiber)
        '409':
          description: Kürzel (slug) bereits vergeben
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /platform/tenants/{tenantId}/suspend:
    post:
      tags: [platform]
      operationId: suspendTenant
      summary: Studio sperren – Studio-Benutzer haben danach keinen Zugriff mehr
      parameters:
        - $ref: '#/components/parameters/TenantId'
      responses:
        '200':
          description: Studio gesperrt (auch wenn es bereits gesperrt war)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TenantResponse'
        '400':
          description: Ungültige Studio-ID
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '401':
          description: Nicht angemeldet
        '403':
          description: Keine Berechtigung (nur Plattform-Betreiber)
        '404':
          description: Studio nicht gefunden
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /platform/tenants/{tenantId}/reactivate:
    post:
      tags: [platform]
      operationId: reactivateTenant
      summary: Gesperrtes Studio wieder freischalten
      parameters:
        - $ref: '#/components/parameters/TenantId'
      responses:
        '200':
          description: Studio aktiv (auch wenn es bereits aktiv war)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TenantResponse'
        '400':
          description: Ungültige Studio-ID
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '401':
          description: Nicht angemeldet
        '403':
          description: Keine Berechtigung (nur Plattform-Betreiber)
        '404':
          description: Studio nicht gefunden
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /platform/tenants/{tenantId}/onboarding/retry:
    post:
      tags: [platform]
      operationId: retryTenantOnboarding
      summary: Fehlgeschlagenes Onboarding sofort erneut versuchen (statt auf die automatische Wiederholung zu warten)
      description: |
        Stellt die fehlgeschlagene Zustellung des Registrierungs-Events für dieses Studio erneut zu. Das Onboarding
        läuft im Hintergrund; das Ergebnis steht danach in onboardingStatus bzw. onboardingError.
        Ohne fehlgeschlagenen Versuch (z. B. bereits abgeschlossen) passiert nichts.
      parameters:
        - $ref: '#/components/parameters/TenantId'
      responses:
        '202':
          description: Erneuter Versuch angestoßen (bzw. nichts zu tun)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TenantResponse'
        '400':
          description: Ungültige Studio-ID
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '401':
          description: Nicht angemeldet
        '403':
          description: Keine Berechtigung (nur Plattform-Betreiber)
        '404':
          description: Studio nicht gefunden
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
components:
  parameters:
    TenantId:
      name: tenantId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreateTenantRequest:
      type: object
      required: [slug, name, adminEmail]
      properties:
        slug:
          type: string
          description: Eindeutiges Kürzel, als Subdomain verwendbar
          pattern: '^[a-z0-9][a-z0-9-]{1,61}[a-z0-9]$'
          example: studio-mueller
        name:
          type: string
          minLength: 1
          maxLength: 200
          example: Fotostudio Müller
        adminEmail:
          type: string
          format: email
          maxLength: 254
          description: E-Mail des ersten Studio-Admins; erhält eine Einladung zum Setzen des Passworts
          example: inhaber@studio-mueller.de
        adminFirstName:
          type: string
          maxLength: 100
          example: Maria
        adminLastName:
          type: string
          maxLength: 100
          example: Müller
    TenantResponse:
      type: object
      required: [id, slug, name, status, onboardingStatus, createdAt]
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
        name:
          type: string
        status:
          type: string
          enum: [ACTIVE, SUSPENDED]
        onboardingStatus:
          type: string
          description: |
            PENDING: Keycloak-Organisation und erster Studio-Admin werden noch angelegt (bei Fehlern automatisch wiederholt).
            COMPLETED: Organisation angelegt, Einladung an den Studio-Admin verschickt.
          enum: [PENDING, COMPLETED]
        onboardingError:
          type: string
          description: |
            Grund des letzten fehlgeschlagenen Onboarding-Versuchs (deutsch, für den Betreiber), z. B. eine E-Mail-Adresse,
            die schon zu einem anderen Studio gehört. Fehlt, solange kein Versuch fehlgeschlagen ist und nach Erfolg.
        onboardingFailedAt:
          type: string
          format: date-time
          description: Zeitpunkt des letzten fehlgeschlagenen Onboarding-Versuchs (nur zusammen mit onboardingError)
        createdAt:
          type: string
          format: date-time
    ProblemDetail:
      description: Fehlerbeschreibung nach RFC 9457
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
...